.Scripts/requirements.txt
basemap==1.4.1
basemap-data==1.3.2
beautifulsoup4==4.13.4
certifi==2025.4.26
cffi==1.17.1
cftime==1.6.4.post1
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
deepdiff==8.5.0
fonttools==4.58.0
idna==3.10
jax==0.6.0
jaxlib==0.6.0
kiwisolver==1.4.8
matplotlib==3.8.4
ml_dtypes==0.5.1
netCDF4==1.7.2
numpy==1.26.4
opt_einsum==3.4.0
orderly-set==5.4.1
packaging==23.2
pillow==11.2.1
pycparser==2.22
pyfar==0.7.2
pyparsing==3.2.3
pyproj==3.6.1
pyshp==2.3.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
scipy==1.15.3
six==1.17.0
sofar==1.2.2
soundfile==0.13.1
soupsieve==2.7
typing_extensions==4.13.2
urllib3==2.4.0